requirements.txt
beautifulsoup4==4.10.0
python-telegram-bot==13.11
pyyaml==6.0
requests==2.27.1

